*dao membership pallet related changes: - removed membership prime - updated unit tests for dao membership pallet - code cleanup
